+++ html_handler.py
@@ -1,4 +1,6 @@
 import asyncio
+import aiohttp
+import json
 from bs4 import BeautifulSoup
 import cssutils
 import logging
